pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autoreduce"
version = "0.3.0"
description = "Python based automated model reduction tools for SBML models"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "BSD-3-Clause"}
authors = [
    {name = "Ayush Pandey", email = "ayushpandey@ucmerced.edu"}
]
keywords = ["SBML", "Automated Model Reduction", "Modeling", "QSSA", "Hill functions"]
dependencies = [
    "python-libsbml",
    "sympy",
    "scipy",
    "numpy",
]

[project.optional-dependencies]
all = [
    "matplotlib",
    "seaborn",
]

[project.urls]
Documentation = "https://autoreduce.readthedocs.io"
Source = "https://github.com/ayush9pandey/AutoReduce"
Tracker = "https://github.com/ayush9pandey/AutoReduce/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=autoreduce --cov-report=xml"

[tool.coverage.run]
source = ["autoreduce"]
omit = ["tests/*", "setup.py"]

[tool.flake8]
max-line-length = 100
exclude = ["build", "dist", ".git", "__pycache__"]
ignore = ["E203", "W503", "E501", "F401"] 